requirements.txt
langgraph>=0.0.51
langchain>=0.1.9
langsmith>=0.0.27
google-generativeai>=0.3.2
langchain-google-genai>=0.0.7
python-dotenv>=1.0.0
playwright>=1.38.0
beautifulsoup4>=4.12.3
tenacity>=8.2.3
pytest>=7.4.4 